# Asp .NET Core Mentoring.UnitTest.API/Models/EmployeeTerritories.cs
// <auto-generated>
// Code generated by Microsoft (R) AutoRest Code Generator.
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.
// </auto-generated>

namespace Swagger.Models
{
    using Newtonsoft.Json;
    using System.Linq;

    public partial class EmployeeTerritories
    {
        /// <summary>
        /// Initializes a new instance of the EmployeeTerritories class.
        /// </summary>
        public EmployeeTerritories()
        {
            CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the EmployeeTerritories class.
        /// </summary>
        public EmployeeTerritories(int? employeeId = default(int?), string territoryId = default(string), Employees employee = default(Employees), Territories territory = default(Territories))
        {
            EmployeeId = employeeId;
            TerritoryId = territoryId;
            Employee = employee;
            Territory = territory;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "employeeId")]
        public int? EmployeeId { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "territoryId")]
        public string TerritoryId { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "employee")]
        public Employees Employee { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "territory")]
        public Territories Territory { get; set; }

    }
}