# Asp .NET Core Mentoring.UnitTest.API/Models/Territories.cs
// <auto-generated>
// Code generated by Microsoft (R) AutoRest Code Generator.
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.
// </auto-generated>

namespace Swagger.Models
{
    using Newtonsoft.Json;
    using System.Collections;
    using System.Collections.Generic;
    using System.Linq;

    public partial class Territories
    {
        /// <summary>
        /// Initializes a new instance of the Territories class.
        /// </summary>
        public Territories()
        {
            CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the Territories class.
        /// </summary>
        public Territories(string territoryId = default(string), string territoryDescription = default(string), int? regionId = default(int?), Region region = default(Region), IList<EmployeeTerritories> employeeTerritories = default(IList<EmployeeTerritories>))
        {
            TerritoryId = territoryId;
            TerritoryDescription = territoryDescription;
            RegionId = regionId;
            Region = region;
            EmployeeTerritories = employeeTerritories;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "territoryId")]
        public string TerritoryId { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "territoryDescription")]
        public string TerritoryDescription { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "regionId")]
        public int? RegionId { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "region")]
        public Region Region { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "employeeTerritories")]
        public IList<EmployeeTerritories> EmployeeTerritories { get; set; }

    }
}