# AutoRestTests/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "version": "v1"
  },
  "paths": {
    "/attributesRoutingImage/{id}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "operationId": "GetCategoryImageViaAttributeRouting",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "operationId": "Get",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Categories/{id}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "operationId": "Get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Categories/UploadSingleFile": {
      "post": {
        "tags": [
          "Categories"
        ],
        "operationId": "Post",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "file",
            "type": "string",
            "format": "binary"
          },
          {
            "in": "formData",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Products": {
      "get": {
        "tags": [
          "Products"
        ],
        "operationId": "Get",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "operationId": "Get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "operationId": "Put",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Products"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "operationId": "Delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Products/CreateNewProduct": {
      "post": {
        "tags": [
          "Products"
        ],
        "operationId": "Post",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Products"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Categories": {
      "type": "object",
      "properties": {
        "categoryId": {
          "format": "int32",
          "type": "integer"
        },
        "categoryName": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "picture": {
          "format": "byte",
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Products"
          }
        }
      }
    },
    "Suppliers": {
      "type": "object",
      "properties": {
        "supplierId": {
          "format": "int32",
          "type": "integer"
        },
        "companyName": {
          "type": "string"
        },
        "contactName": {
          "type": "string"
        },
        "contactTitle": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "region": {
          "type": "string"
        },
        "postalCode": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "fax": {
          "type": "string"
        },
        "homePage": {
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Products"
          }
        }
      }
    },
    "CustomerDemographics": {
      "type": "object",
      "properties": {
        "customerTypeId": {
          "type": "string"
        },
        "customerDesc": {
          "type": "string"
        },
        "customerCustomerDemo": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CustomerCustomerDemo"
          }
        }
      }
    },
    "CustomerCustomerDemo": {
      "type": "object",
      "properties": {
        "customerId": {
          "type": "string"
        },
        "customerTypeId": {
          "type": "string"
        },
        "customer": {
          "$ref": "#/definitions/Customers"
        },
        "customerType": {
          "$ref": "#/definitions/CustomerDemographics"
        }
      }
    },
    "Customers": {
      "type": "object",
      "properties": {
        "customerId": {
          "type": "string"
        },
        "companyName": {
          "type": "string"
        },
        "contactName": {
          "type": "string"
        },
        "contactTitle": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "region": {
          "type": "string"
        },
        "postalCode": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "fax": {
          "type": "string"
        },
        "customerCustomerDemo": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CustomerCustomerDemo"
          }
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Orders"
          }
        }
      }
    },
    "Region": {
      "type": "object",
      "properties": {
        "regionId": {
          "format": "int32",
          "type": "integer"
        },
        "regionDescription": {
          "type": "string"
        },
        "territories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Territories"
          }
        }
      }
    },
    "Territories": {
      "type": "object",
      "properties": {
        "territoryId": {
          "type": "string"
        },
        "territoryDescription": {
          "type": "string"
        },
        "regionId": {
          "format": "int32",
          "type": "integer"
        },
        "region": {
          "$ref": "#/definitions/Region"
        },
        "employeeTerritories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EmployeeTerritories"
          }
        }
      }
    },
    "EmployeeTerritories": {
      "type": "object",
      "properties": {
        "employeeId": {
          "format": "int32",
          "type": "integer"
        },
        "territoryId": {
          "type": "string"
        },
        "employee": {
          "$ref": "#/definitions/Employees"
        },
        "territory": {
          "$ref": "#/definitions/Territories"
        }
      }
    },
    "Employees": {
      "type": "object",
      "properties": {
        "employeeId": {
          "format": "int32",
          "type": "integer"
        },
        "lastName": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "titleOfCourtesy": {
          "type": "string"
        },
        "birthDate": {
          "format": "date-time",
          "type": "string"
        },
        "hireDate": {
          "format": "date-time",
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "region": {
          "type": "string"
        },
        "postalCode": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "homePhone": {
          "type": "string"
        },
        "extension": {
          "type": "string"
        },
        "photo": {
          "format": "byte",
          "type": "string"
        },
        "notes": {
          "type": "string"
        },
        "reportsTo": {
          "format": "int32",
          "type": "integer"
        },
        "photoPath": {
          "type": "string"
        },
        "reportsToNavigation": {
          "$ref": "#/definitions/Employees"
        },
        "employeeTerritories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EmployeeTerritories"
          }
        },
        "inverseReportsToNavigation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Employees"
          }
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Orders"
          }
        }
      }
    },
    "Shippers": {
      "type": "object",
      "properties": {
        "shipperId": {
          "format": "int32",
          "type": "integer"
        },
        "companyName": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Orders"
          }
        }
      }
    },
    "Orders": {
      "type": "object",
      "properties": {
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "customerId": {
          "type": "string"
        },
        "employeeId": {
          "format": "int32",
          "type": "integer"
        },
        "orderDate": {
          "format": "date-time",
          "type": "string"
        },
        "requiredDate": {
          "format": "date-time",
          "type": "string"
        },
        "shippedDate": {
          "format": "date-time",
          "type": "string"
        },
        "shipVia": {
          "format": "int32",
          "type": "integer"
        },
        "freight": {
          "format": "double",
          "type": "number"
        },
        "shipName": {
          "type": "string"
        },
        "shipAddress": {
          "type": "string"
        },
        "shipCity": {
          "type": "string"
        },
        "shipRegion": {
          "type": "string"
        },
        "shipPostalCode": {
          "type": "string"
        },
        "shipCountry": {
          "type": "string"
        },
        "customer": {
          "$ref": "#/definitions/Customers"
        },
        "employee": {
          "$ref": "#/definitions/Employees"
        },
        "shipViaNavigation": {
          "$ref": "#/definitions/Shippers"
        },
        "orderDetails": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderDetails"
          }
        }
      }
    },
    "OrderDetails": {
      "type": "object",
      "properties": {
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "unitPrice": {
          "format": "double",
          "type": "number"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "discount": {
          "format": "float",
          "type": "number"
        },
        "order": {
          "$ref": "#/definitions/Orders"
        },
        "product": {
          "$ref": "#/definitions/Products"
        }
      }
    },
    "Products": {
      "required": [
        "productName",
        "quantityPerUnit"
      ],
      "type": "object",
      "properties": {
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "productName": {
          "maxLength": 80,
          "minLength": 0,
          "type": "string"
        },
        "supplierId": {
          "format": "int32",
          "type": "integer"
        },
        "categoryId": {
          "format": "int32",
          "type": "integer"
        },
        "quantityPerUnit": {
          "maxLength": 80,
          "minLength": 0,
          "type": "string"
        },
        "unitPrice": {
          "format": "double",
          "type": "number"
        },
        "unitsInStock": {
          "format": "int32",
          "type": "integer"
        },
        "unitsOnOrder": {
          "format": "int32",
          "type": "integer"
        },
        "reorderLevel": {
          "format": "int32",
          "type": "integer"
        },
        "discontinued": {
          "type": "boolean"
        },
        "category": {
          "$ref": "#/definitions/Categories"
        },
        "supplier": {
          "$ref": "#/definitions/Suppliers"
        },
        "orderDetails": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderDetails"
          }
        }
      }
    }
  }
}